Refactored YaraEngine actor messages to own module
